refactor(utils): move write_tempfile to FS
