+++ src/util/debug.lua
@@ -1,10 +1,8 @@
 --- @diagnostic disable: redefined-local
 
-require("util.filesystem")
 require("util.string")
 require("util.table")
 local tc = require("util.termcolor")
-local OS = require("util.os")
 
 local tab = '  '
 
@@ -402,35 +400,6 @@ Debug = {
     end
   end,
 
-  --- @param content str
-  --- @param ext string?
-  --- @param fixname string?
-  write_tempfile = function(content, ext, fixname)
-    local function create_temp()
-      local cmd = 'mktemp -u -p .'
-      if string.is_non_empty_string(ext) then
-        cmd = string.format('%s --suffix .%s', cmd, ext)
-      end
-      local _, result = OS.runcmd(cmd)
-      return result
-    end
-    local name =
-        string.is_non_empty_string(fixname)
-        and fixname .. (ext and '.' .. ext or '')
-        or create_temp()
-    local mok, merr = FS.mkdirp('./.debug')
-    if not mok then
-      return false, merr
-    end
-    local path = FS.join_path('./.debug', name)
-
-    local data = string.unlines(content)
-    local ok, err = FS.write(path, data)
-    if not ok then
-      return false, err
-    end
-    return ok
-  end,
 }
 
 local printer = (function()
