+++ tests/interpreter/analyzer_spec.lua
@@ -1,6 +1,7 @@
 local parser = require("model.lang.lua.parser")('metalua')
 local analyzer = require("model.lang.lua.analyze")
 local term = require("util.termcolor")
+local FS = require("util.filesystem")
 require("util.color")
 require("util.debug")
 
@@ -80,7 +81,7 @@ describe('analyzer #analyzer', function()
                     tree
                   )
                   local wres, werr =
-                      Debug.write_tempfile(f, 'json5', fn)
+                      FS.write_tempfile(f, 'json5', fn)
                   if not wres then
                     Log.warn("no write", werr)
                   end
